+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ New features and notable changes:
 
 - Add support for colored logging. (:issue:`887`)
 - Add support for TOML configuration format. (:issue:`881`) 
+- Add support for Clover XML output format. (:issue:`888`) 
 
 Bug fixes and small improvements: